refactor greek
