+++ .pylintrc
@@ -60,7 +60,8 @@ confidence=
 # --enable=similarities". If you want to run only the classes checker, but have
 # no Warning level messages displayed, use "--disable=all --enable=classes
 # --disable=W".
-disable=print-statement,
+disable=bad-continuation,
+        print-statement,
         parameter-unpacking,
         unpacking-in-except,
         old-raise-syntax,
