+++ src/cltkv1/alphabet/__init__.py
@@ -1,7 +1,6 @@
 """Modules for accessing the alphabets and character sets of in-scope CLTK languages.
 
 TODO: Add punjabi? if so west or east? west2386/panj1256
-TODO: Decide where/how to put cltk/corpus/sanskrit/itrans
 TODO: Add telugu
 TODO: Add Urdu
 """
