+++ src/cltkv1/alphabet/grc/__init__.py
@@ -1,0 +1,6 @@
+"""Init for Greek language alphabet and encoding tools. Import `grc` alphabet
+module so that users may import it according to the API used for others
+(e.g., ``from cltkv1.alphabet import grc``).
+"""
+
+from .grc import *
